Scripts within the popup must not be inline (CSP)
Finally discovered why my inline script wasn't working and how to debug!
You can right mouse click on the toolbar icon and "inspect", which shows
an error in the console that explains how inline scripts are not
permitted by default.
Extracting the script out in to a separate file seems to be fine though.

As for communicating between background.js and the popup, they do seem
to share localStorage, but this post shows how to send and listen for
messages between them:
  http://stackoverflow.com/questions/13546778/how-to-communicate-between-popup-js-and-background-js-in-chrome-extension