Use localStorage to store the result of the builds
This effectively caches the result of the builds, so now when the popup
receives a new message from the background worker, it is just a signal
to update the UI based on the cache.

This means during the refresh or if the internet goes away, there will
be the only builds still visible, they'll just be a little out of date.