Make the refreshRate user configurable
Changing the config from the UI needs a way to reset the interval thats
already running, so now the background process can receive a message
that will signal it to reload initialize itself with new configuration
from localStorage.

Specifically the refreshRate as the others are retrieved from localStore
every time updateBuilds is called.