Use mocha to do tests (needs to be modularised)
I just wanted to write unit tests to help me write the overallState
function, so I got the mocha+chai but couldn't import the function
because I'm not used modules.

Eventually I'd like to modularise this code and when I do, I will
probably want to have this test as part of a suite, so I'm putting this
on a branch for now.

Not the greatest practice I know :P