Use a gear logo instead of the word "Configure"
This gear logo was downloaded from:
  https://raw.githubusercontent.com/encharm/Font-Awesome-SVG-PNG

Which is derived from:
  http://fontawesome.io/

I didn't think I needed the entire icon font for just one icon.
I am considering using an icon for the git-branch logo, if I end up
using more icons, it may be easier to just use FA as it was intended.